Replace CLI calls in venv with API calls
The first version of this tool was meant to be a single file script
without external dependencies. This was achieved by creating a
virtual environment with 'venv' and populate it with the necessary
tools. This approach works well and is still valid.

The tool now outgrew the one-single-file phase and thus the
independence from external dependencies is not a constraint anymore.