Add support for 'output_file_win' setting
On Windows ('nt' platform) it is convenient to add the prefix '.pyz'
to the file names of the resulting zipapp archives, since on Windows
the 'pyz' extension is commonly added to the 'PATHEXT' environment
variable. Files with such an extension are then recognized as
executables and it is then possible for a zipapp 'something.pyz' to
be called by leaving out the extension and simply typing 'something'.

On 'posix' platforms the 'executable' file mode bit is used to
indicate which files are executable and thus the file name suffix is
not relevant and often it is preferably left out entirely.